+++ client/src/Craig.js
@@ -12,7 +12,7 @@ import {
   ReleaseNotes,
   Summary,
   Projects,
-  ToggleFormPage
+  ToggleFormPage,
 } from "./components";
 import { invalidForms, state } from "./lib";
 import { default as constants } from "./lib/constants";
@@ -21,7 +21,7 @@ import { JsonDocs } from "./components/pages/JsonDocs";
 import Tutorial from "./components/pages/tutorial/Tutorial";
 import { notificationText } from "./lib/forms/utils";
 
-const withRouter = Page => props => {
+const withRouter = (Page) => (props) => {
   const params = useParams();
   return <Page {...props} params={params} />;
 };
@@ -52,7 +52,7 @@ class Craig extends React.Component {
         storeName: storeName,
         projects: JSON.parse(projectInStorage),
         store: craig.store,
-        visited: window.localStorage.getItem("craigVisited")
+        visited: window.localStorage.getItem("craigVisited"),
       };
     } catch (err) {
       window.location.pathname = "/resetState";
@@ -93,11 +93,11 @@ class Craig extends React.Component {
       title: "Success",
       kind: "success",
       text: message || `Successfully updated ${updatedForm}`,
-      timeout: 3000
+      timeout: 3000,
     };
     this.setState(
       {
-        store: craig.store
+        store: craig.store,
       },
       () => {
         this.notify(notification);
@@ -110,7 +110,7 @@ class Craig extends React.Component {
       title: "Error",
       kind: "error",
       text: "An unexpected error has occurred.",
-      timeout: 3000
+      timeout: 3000,
     };
     this.notify(notification);
   }
@@ -138,8 +138,8 @@ class Craig extends React.Component {
    * @param {*} notification
    */
   notify(notification) {
-    this.setState(prevState => ({
-      notifications: [...prevState.notifications, notification]
+    this.setState((prevState) => ({
+      notifications: [...prevState.notifications, notification],
     }));
   }
 
@@ -156,7 +156,7 @@ class Craig extends React.Component {
     return {
       name: "",
       description: "",
-      json: new state().store.json
+      json: new state().store.json,
     };
   }
 
@@ -175,7 +175,7 @@ class Craig extends React.Component {
       name: stateData.name,
       description: stateData.description,
       json: stateData.json,
-      last_save: now
+      last_save: now,
     };
 
     // if the project name is changing, remove the old key from projects object