+++ client/src/components/utils/NoEdgeNetworkTile.js
@@ -6,12 +6,18 @@ import "./no-secrets-manager-tile.css";
 export const NoEdgeNetworkTile = () => {
   return (
     <Tile className="tileBackground displayFlex alignItemsCenter wrap marginTop">
-      <CloudAlerting size="24" className="iconMargin" /> No Edge Network. Go
-      back to the{" "}
-      <a className="no-secrets-link" href="/">
-        Home page
-      </a>{" "}
-      to enable Edge Networking.
+      <span>
+        <CloudAlerting size="24" className="iconMargin" /> No Edge Network. Go
+        back to the{" "}
+        <a className="no-secrets-link" href="/">
+          Home page
+        </a>{" "}
+        to enable Edge Networking.{" "}
+        <em>
+          Dynamic Scalable Subnets must be disabled to create an Edge VPC
+          network.
+        </em>
+      </span>
     </Tile>
   );
 };