+++ client/src/lib/forms/index.js
@@ -12,7 +12,7 @@ const {
   invalidCidrText,
   invalidCbrRuleText,
   invalidCbrZoneText,
-  invalidProjectNameText
+  invalidProjectNameText,
 } = require("./text-callbacks");
 const {
   invalidName,
@@ -37,7 +37,7 @@ const {
   invalidCbrRule,
   invalidCbrZone,
   validRecord,
-  invalidDnsZoneName
+  invalidDnsZoneName,
 } = require("./invalid-callbacks");
 const { propsMatchState } = require("./props-match-state");
 const {
@@ -45,7 +45,7 @@ const {
   invalidPort,
   forceShowForm,
   disableSshKeyDelete,
-  invalidCidrBlock
+  invalidCidrBlock,
 } = require("./disable-save");
 const { hasDuplicateName } = require("./duplicate-name");
 const { getSubnetTierStateData, getTierSubnets } = require("./state-data");
@@ -53,7 +53,7 @@ const {
   formatConfig,
   copyAclModalContent,
   copyRuleCodeMirrorData,
-  copySgModalContent
+  copySgModalContent,
 } = require("./format-json");
 const {
   setFormRgList,
@@ -62,7 +62,7 @@ const {
   setFormEncryptionKeyList,
   setFormSubnetList,
   setDeleteDisabledMessage,
-  setFormSgList
+  setFormSgList,
 } = require("./form-props");
 const { leftNavItemClassName } = require("./class-names");
 const { notificationText } = require("./utils");
@@ -125,5 +125,5 @@ module.exports = {
   setFormSgList,
   invalidProjectNameText,
   validRecord,
-  invalidDnsZoneName
+  invalidDnsZoneName,
 };